third_party/blink/web_tests/virtual/customizable-select-disabled/fast/forms/select/customizable-select/allowed-select-descendants-console-message-expected.txt

CONSOLE WARNING: A button tag was parsed inside of a <select> which was not inserted into the document. This is not valid HTML and the behavior may be changed in future versions of chrome.
CONSOLE WARNING: A div tag was parsed inside of a <select> which was not inserted into the document. This is not valid HTML and the behavior may be changed in future versions of chrome.
CONSOLE WARNING: A noscript tag was parsed inside of a <select> which was not inserted into the document. This is not valid HTML and the behavior may be changed in future versions of chrome.
 
..
..
 
..
 
..
..
   
